Move Handoff update structs into lib
